// src/entities/User/model/types/UserSchema.ts
import { UserRole } from "../consts/consts";
import { FeatureFlags } from "@/shared/types/featureFlags";

export interface User {
  id: string;
  username: string;
  avatar?: string;
  roles?: UserRole[];
  features?: FeatureFlags;
}

export interface UserSchema {
  authData?: User;

  mounted: boolean;
}